refactor: extracts tool and mcp related calls to playground-ui
