+++ .changeset/tiny-drinks-play.md
@@ -1,0 +1,5 @@
+---
+'@mastra/playground-ui': patch
+---
+
+Move some components to playground-ui for usage in cloud
