+++ packages/playground-ui/src/components/ui/copy-button.tsx
@@ -1,35 +1,41 @@
 'use client';
 
-import { Check, Copy } from 'lucide-react';
+import { CopyIcon } from 'lucide-react';
 
-import { Button } from '../../components/ui/button';
 import { useCopyToClipboard } from '../../hooks/use-copy-to-clipboard';
-import { cn } from '../../lib/utils';
+
+import { Tooltip, TooltipContent, TooltipTrigger } from './tooltip';
+import { Icon, IconProps } from '@/ds/icons';
 
 type CopyButtonProps = {
   content: string;
   copyMessage?: string;
+  tooltip?: string;
   className?: string;
+  iconSize?: IconProps['size'];
 };
 
-export function CopyButton({ content, copyMessage, className }: CopyButtonProps) {
-  const { isCopied, handleCopy } = useCopyToClipboard({
+export function CopyButton({
+  content,
+  copyMessage,
+  tooltip = 'Copy to clipboard',
+  iconSize = 'default',
+}: CopyButtonProps) {
+  const { handleCopy } = useCopyToClipboard({
     text: content,
     copyMessage,
   });
 
   return (
-    <Button
-      variant="ghost"
-      size="icon"
-      className={cn('relative h-6 w-6', className)}
-      aria-label="Copy to clipboard"
-      onClick={handleCopy}
-    >
-      <div className="absolute inset-0 flex items-center justify-center">
-        <Check className={cn('h-4 w-4 transition-transform ease-in-out', isCopied ? 'scale-100' : 'scale-0')} />
-      </div>
-      <Copy className={cn('h-4 w-4 transition-transform ease-in-out', isCopied ? 'scale-0' : 'scale-100')} />
-    </Button>
+    <Tooltip>
+      <TooltipTrigger asChild>
+        <button onClick={handleCopy} type="button">
+          <Icon className="transition-colors hover:bg-surface4 rounded-lg text-icon3 hover:text-icon6" size={iconSize}>
+            <CopyIcon />
+          </Icon>
+        </button>
+      </TooltipTrigger>
+      <TooltipContent>{tooltip}</TooltipContent>
+    </Tooltip>
   );
 }
