+++ packages/playground-ui/src/domains/agents/index.tsx
@@ -12,3 +12,4 @@ export * from './utils/extractPrompt';
 export * from './hooks/use-agents';
 export * from './hooks/use-agent';
 export * from './hooks/use-execute-agent-tool';
+export * from './components/AgentToolPanel';
