+++ packages/playground-ui/src/domains/mcps/index.ts
@@ -1,2 +1,4 @@
 export * from './components/mcp-table/mcp-table';
-export * from './hooks/useMCPServerTools';
+export * from './hooks/index';
+export * from './components/MCPToolPanel';
+export * from './components/MCPDetail';
