+++ packages/playground-ui/src/domains/tools/components/ToolExecutor.tsx
@@ -1,10 +1,12 @@
-import { DynamicForm, MainContentContent, ToolInformation } from '@mastra/playground-ui';
+import { DynamicForm } from '@/components/dynamic-form';
 import { CopyButton } from '@/components/ui/copy-button';
 import { ZodType } from 'zod';
+import { ToolInformation } from '@/domains/tools/components/ToolInformation';
 import { jsonLanguage } from '@codemirror/lang-json';
 import { useCodemirrorTheme } from '@/components/syntax-highlighter';
 import CodeMirror from '@uiw/react-codemirror';
 import { MCPToolType } from '@mastra/core/mcp';
+import { MainContentContent } from '@/components/ui/containers/MainContent';
 
 export interface ToolExecutorProps {
   isExecutingTool: boolean;
