+++ packages/playground-ui/src/domains/tools/components/ToolIcon.tsx
@@ -1,0 +1,9 @@
+import { AgentIcon } from '@/ds/icons/AgentIcon';
+import { ToolsIcon } from '@/ds/icons/ToolsIcon';
+import { WorkflowIcon } from '@/ds/icons/WorkflowIcon';
+
+export const ToolIconMap = {
+  agent: AgentIcon,
+  workflow: WorkflowIcon,
+  tool: ToolsIcon,
+};
