+++ packages/playground-ui/src/domains/tools/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './use-execute-tool';
+export * from './use-all-tools';
