+++ packages/playground-ui/src/domains/tools/hooks/use-all-tools.ts
@@ -1,0 +1,22 @@
+import { usePlaygroundStore } from '@/store/playground-store';
+import { useQuery } from '@tanstack/react-query';
+import { useMastraClient } from '@mastra/react';
+
+export const useTools = () => {
+  const { runtimeContext } = usePlaygroundStore();
+  const client = useMastraClient();
+  return useQuery({
+    queryKey: ['tools'],
+    queryFn: () => client.getTools(runtimeContext),
+  });
+};
+
+export const useTool = (toolId: string) => {
+  const client = useMastraClient();
+  const { runtimeContext } = usePlaygroundStore();
+
+  return useQuery({
+    queryKey: ['tool', toolId],
+    queryFn: () => client.getTool(toolId).details(runtimeContext),
+  });
+};
