+++ packages/playground-ui/src/domains/tools/index.ts
@@ -1,2 +1,3 @@
 export * from './components/tool-table/tool-table';
+export * from './components';
 export * from './hooks';
