+++ packages/playground-ui/src/lib/framework.tsx
@@ -32,7 +32,7 @@ type LinkComponentPaths = {
   toolLink: (toolId: string) => string;
 
   mcpServerLink: (serverId: string) => string;
-
+  mcpServerToolLink: (serverId: string, toolId: string) => string;
   workflowRunLink: (workflowId: string, runId: string) => string;
 };
 
@@ -57,6 +57,7 @@ const LinkComponentContext = createContext<{
     scorerLink: () => '',
     toolLink: () => '',
     mcpServerLink: () => '',
+    mcpServerToolLink: () => '',
     workflowRunLink: () => '',
   },
 });
