+++ packages/playground/src/pages/tools/agent-tool/index.tsx
@@ -1,55 +1,20 @@
-import jsonSchemaToZod from 'json-schema-to-zod';
-import { useState } from 'react';
 import { Link, useParams } from 'react-router';
-import { parse } from 'superjson';
-import { z } from 'zod';
 
-import { resolveSerializedZodOutput } from '@/components/dynamic-form/utils';
-import { useAgent, useExecuteAgentTool } from '@mastra/playground-ui';
-
-import ToolExecutor from '../tool-executor';
 import {
   Header,
   Crumb,
   Breadcrumb,
-  usePlaygroundStore,
-  Txt,
   MainContentLayout,
   AgentIcon,
   Icon,
   HeaderAction,
   Button,
   DocsIcon,
+  AgentToolPanel,
 } from '@mastra/playground-ui';
 
 const AgentTool = () => {
   const { toolId, agentId } = useParams();
-  const { runtimeContext: playgroundRuntimeContext } = usePlaygroundStore();
-
-  const { mutateAsync: executeTool, isPending: isExecutingTool, error } = useExecuteAgentTool();
-  const [result, setResult] = useState<any>(null);
-
-  const { data: agent, isLoading: isAgentLoading } = useAgent(agentId!);
-
-  const tool = Object.values(agent?.tools ?? {}).find(tool => tool.id === toolId);
-
-  const handleExecuteTool = async (data: any) => {
-    if (!agent || !tool) return;
-
-    const result = await executeTool({
-      agentId: agentId!,
-      toolId: tool.id,
-      input: data,
-      playgroundRuntimeContext,
-    });
-    setResult(result);
-  };
-
-  const zodInputSchema = tool?.inputSchema
-    ? resolveSerializedZodOutput(jsonSchemaToZod(parse(tool?.inputSchema)))
-    : z.object({});
-
-  const shouldShowEmpty = !agent || !tool;
 
   return (
     <MainContentLayout>
@@ -79,23 +44,7 @@ const AgentTool = () => {
         </HeaderAction>
       </Header>
 
-      {isAgentLoading ? null : shouldShowEmpty ? (
-        <div className="py-12 text-center px-6">
-          <Txt variant="header-md" className="text-icon3">
-            Agent or tool not found
-          </Txt>
-        </div>
-      ) : (
-        <ToolExecutor
-          executionResult={result}
-          errorString={error?.message ?? ''}
-          isExecutingTool={isExecutingTool}
-          zodInputSchema={zodInputSchema}
-          handleExecuteTool={handleExecuteTool}
-          toolDescription={tool.description}
-          toolId={tool.id}
-        />
-      )}
+      <AgentToolPanel toolId={toolId!} agentId={agentId!} />
     </MainContentLayout>
   );
 };
