+++ packages/playground/src/pages/tools/tool/index.tsx
@@ -1,48 +1,19 @@
-import { jsonSchemaToZod } from 'json-schema-to-zod';
-
 import { Link, useParams } from 'react-router';
-import { parse } from 'superjson';
-import { z } from 'zod';
 
-import { resolveSerializedZodOutput } from '@/components/dynamic-form/utils';
-
-import { useTool } from '@/hooks/use-all-tools';
-import { useExecuteTool } from '@mastra/playground-ui';
-
-import ToolExecutor from '../tool-executor';
 import {
   Header,
   Breadcrumb,
   Crumb,
-  usePlaygroundStore,
-  Txt,
   Icon,
   ToolsIcon,
   HeaderAction,
   Button,
   DocsIcon,
+  ToolPanel,
 } from '@mastra/playground-ui';
 
 const Tool = () => {
   const { toolId } = useParams();
-  const { tool, isLoading } = useTool(toolId!);
-
-  const { mutateAsync: executeTool, isPending: isExecuting, data: result } = useExecuteTool();
-  const { runtimeContext: playgroundRuntimeContext } = usePlaygroundStore();
-
-  const handleExecuteTool = async (data: any) => {
-    if (!tool) return;
-
-    return executeTool({
-      toolId: tool.id,
-      input: data,
-      runtimeContext: playgroundRuntimeContext,
-    });
-  };
-
-  const zodInputSchema = tool?.inputSchema
-    ? resolveSerializedZodOutput(jsonSchemaToZod(parse(tool?.inputSchema)))
-    : z.object({});
 
   return (
     <div className="h-full w-full overflow-y-hidden">
@@ -68,22 +39,8 @@ const Tool = () => {
           </Button>
         </HeaderAction>
       </Header>
-      {isLoading ? null : !tool ? (
-        <div className="py-12 text-center px-6">
-          <Txt variant="header-md" className="text-icon3">
-            Tool not found
-          </Txt>
-        </div>
-      ) : (
-        <ToolExecutor
-          executionResult={result}
-          isExecutingTool={isExecuting}
-          zodInputSchema={zodInputSchema}
-          handleExecuteTool={handleExecuteTool}
-          toolDescription={tool.description}
-          toolId={tool.id}
-        />
-      )}
+
+      <ToolPanel toolId={toolId!} />
     </div>
   );
 };
