+++ packages/playground/src/types.ts
@@ -1,5 +1,3 @@
-import { AgentIcon, ToolsIcon, WorkflowIcon } from '@mastra/playground-ui';
-
 export interface Message {
   id: string;
   role: 'user' | 'assistant';
@@ -15,9 +13,3 @@ export interface ChatProps {
   memory?: boolean;
   url?: string;
 }
-
-export const ToolIconMap = {
-  agent: AgentIcon,
-  workflow: WorkflowIcon,
-  tool: ToolsIcon,
-};
